requirements.txt
requests==2.9.1
Flask_Security==3.0.0
SQLAlchemy==1.1.15
Flask_Login==0.4.1
Flask_WTF==0.14.2
Flask_SQLAlchemy==2.3.2
Flask==0.12.2
WTForms==2.1
Flask_Admin==1.5.1
GitHub_Flask==3.2.0
sqlalchemy_migrate==0.11.0
